requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
flask>=2.0.0
torch>=1.9.0
torchvision>=0.10.0
